+++ CHANGES.rst
@@ -9,6 +9,7 @@ v0.9.0, xxx-xx-xx
   as well as for stochastic independence.
 - Add support for calculating capacity coefficients.
 - Add `tests.find_nearest()`.
+- Officially support Python 3.
 
 ******************
 v0.8.0, 2016-08-29